Go 1.6 introduced `blocks`. Render has removed `blocks` from 1.6, but introduced `partials` as a drop in replacement.
