+++ .travis.yml
@@ -4,6 +4,7 @@ go:
   - 1.3
   - 1.4
   - 1.5
+  - tip
 
 install:
   - go get github.com/eknkc/amber
