+++ README.md
@@ -2,6 +2,9 @@
 
 Render is a package that provides functionality for easily rendering JSON, XML, text, binary data, and HTML templates. This package is based on the [Martini](https://github.com/go-martini/martini) [render](https://github.com/martini-contrib/render) work.
 
+## Block Deprecation Notice
+Go 1.6 introduces a new [block](https://github.com/golang/go/blob/release-branch.go1.6/src/html/template/example_test.go#L128) action. This conflicts with Render's included `block` template function. To provide an easy migration path, a new function was created called `partial`. It is a duplicate of the old `block` function. It is advised that all users of the `block` function update their code to avoid any issues in the future. Previous to Go 1.6, Render's `block` functionality will continue to work but a message will be logged urging you to migrate to the new `partial` function.
+
 ## Usage
 Render can be used with pretty much any web framework providing you can access the `http.ResponseWriter` from your handler. The rendering functions simply wraps Go's existing functionality for marshaling and rendering data.
 
@@ -84,7 +87,7 @@ r := render.New(render.Options{
     AssetNames: func() []string { // Return a list of asset names for the Asset function
       return []string{"filename.tmpl"}
     },
-    Layout: "layout", // Specify a layout template. Layouts can call {{ yield }} to render the current template or {{ block "css" }} to render a block from the current template
+    Layout: "layout", // Specify a layout template. Layouts can call {{ yield }} to render the current template or {{ partial "css" }} to render a partial from the current template.
     Extensions: []string{".tmpl", ".html"}, // Specify extensions to load for templates.
     Funcs: []template.FuncMap{AppHelpers}, // Specify helper function maps for templates to access.
     Delims: render.Delims{"{[{", "}]}"}, // Sets delimiters to the specified strings.
@@ -97,7 +100,7 @@ r := render.New(render.Options{
     IsDevelopment: true, // Render will now recompile the templates on every HTML response.
     UnEscapeHTML: true, // Replace ensure '&<>' are output correctly (JSON only).
     StreamingJSON: true, // Streams the JSON response via json.Encoder.
-    RequireBlocks: true, // Return an error if a template is missing a block used in a layout.
+    RequirePartials: true, // Return an error if a template is missing a partial used in a layout.
     DisableHTTPErrorRendering: true, // Disables automatic rendering of http.StatusInternalServerError when an error occurs.
 })
 // ...
@@ -128,7 +131,7 @@ r := render.New(render.Options{
     IsDevelopment: false,
     UnEscapeHTML: false,
     StreamingJSON: false,
-    RequireBlocks: false,
+    RequirePartials: false,
     DisableHTTPErrorRendering: false,
 })
 ~~~
@@ -164,7 +167,7 @@ You can also load templates from memory by providing the Asset and AssetNames op
 e.g. when generating an asset file using [go-bindata](https://github.com/jteeuwen/go-bindata).
 
 ### Layouts
-Render provides `yield` and `block` functions for layouts to access:
+Render provides `yield` and `partial` functions for layouts to access:
 ~~~ go
 // ...
 r := render.New(render.Options{
@@ -178,16 +181,16 @@ r := render.New(render.Options{
 <html>
   <head>
     <title>My Layout</title>
-    <!-- Render the block template called `css-$current_template` here -->
-    {{ block "css" }}
+    <!-- Render the partial template called `css-$current_template` here -->
+    {{ partial "css" }}
   </head>
   <body>
-    <!-- render the block template called `header-$current_template` here -->
-    {{ block "header" }}
+    <!-- render the partial template called `header-$current_template` here -->
+    {{ partial "header" }}
     <!-- Render the current template here -->
     {{ yield }}
-    <!-- render the block template called `footer-$current_template` here -->
-    {{ block "footer" }}
+    <!-- render the partial template called `footer-$current_template` here -->
+    {{ partial "footer" }}
   </body>
 </html>
 ~~~
@@ -205,8 +208,8 @@ r := render.New(render.Options{
 </html>
 ~~~
 
-Blocks are defined by individual templates as seen below. The block template's
-name needs to be defined as "{block name}-{template name}".
+Partials are defined by individual templates as seen below. The partial template's
+name needs to be defined as "{partial name}-{template name}".
 ~~~ html
 <!-- templates/home.tmpl -->
 {{ define "header-home" }}
@@ -218,9 +221,9 @@ name needs to be defined as "{block name}-{template name}".
 {{ end }}
 ~~~
 
-By default, the template is not required to define all blocks referenced in the
+By default, the template is not required to define all partials referenced in the
 layout. If you want an error to be returned when a template does not define a
-block, set `Options.RequireBlocks = true`.
+partial, set `Options.RequirePartials = true`.
 
 ### Character Encodings
 Render will automatically set the proper Content-Type header based on which function you call. See below for an example of what the default settings would output (note that UTF-8 is the default, and binary data does not output the charset):
