+++ fixtures/partials/content-partial.tmpl
@@ -1,0 +1,1 @@
+{{define "after-content-partial"}}after {{ . }}{{end}}
