+++ fixtures/partials/content.tmpl
@@ -1,0 +1,2 @@
+{{define "before-content"}}before {{ . }}{{end}}
+{{define "after-content"}}after {{ . }}{{end}}
