+++ fixtures/partials/layout.tmpl
@@ -1,0 +1,3 @@
+{{ partial "before" }}
+<h1>during</h1>
+{{ partial "after" }}
